refactor(test.js): create yakbak module
